Use the highest DM we can.
This is in some ways less than ideal, but we want to let people continue
to use 0.10.x if they have to.  Obviously ideally people would upgrade
to 1.x quickly.

While this loose dep does have some risk, regarding DM 2.0, we only use
a very small amount of DM's functionality. It's only from the public (and
widely used) API and is unlikely to change between versions.